:truck: :twisted_rightwards_arrows: Input system
