+++ Disappear/Assets/Multiplayer/Scripts/MultiplayerManager.cs
@@ -11,6 +11,7 @@ using UnityEngine.PlayerLoop;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
 using WaG;
+using WAG.Core.Controls;
 using Hashtable = ExitGames.Client.Photon.Hashtable;
 
 public class MultiplayerManager : MonoBehaviourPunCallbacks
